requirements.txt
# GUI Framework
PyQt6==6.7.0
PyQt6-Qt6==6.7.0

# Video Processing
opencv-python==4.8.1.78
moviepy==1.0.3
imageio==2.31.1
imageio-ffmpeg==0.4.9

# Image Processing
Pillow==10.0.0
numpy==1.24.3

# File Handling
send2trash==1.8.2

# Performance
numba==0.57.1

# Packaging
pyinstaller==5.13.0
auto-py-to-exe==2.38.0

# Development
pytest==7.4.0
black==23.7.0
flake8==6.0.0